server/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "python-jose[cryptography]>=3.4.0",
    "pydantic-settings>=2.9.1",
    "requests>=2.32.3",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.3",
    "uvloop>=0.21.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.2.3",
    "pygithub>=2.6.1",
    "alembic>=1.16.1",
    "e2b-code-interpreter==1.2.0b5",
    "types-requests>=2.32.0.20250602",
    "anthropic>=0.40.0",
    "nanoid>=2.0.0",
    "cachetools>=6.0.0",
]

[tool.uv]
prerelease = "allow"


[tool.mypy]
python_version = "3.13"
plugins = ['pydantic.mypy']
ignore_missing_imports = true
strict = true

[tool.ruff]
lint.select = ["E302"]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "types-cachetools>=6.0.0.20250525",
]
